+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # HetrickCV Changelog
 
 ## 2.4.0 (in progress)
-- Added dark mode for all panels! (in progress)
+- Added dark mode for all panels!
 - Add Phasor Humanize module.
+- Add Phasor Mixer module.
 - Add Phasor Swing module.
 - Fix Phasor Randomizer, Phasor Shifter, Phasor Substep Shaper, and Phasor Stutter Steps knobs initializing with a 0 display.